+++ packages/gamut-labs/src/brand/ProLogo/index.tsx
@@ -1,5 +1,6 @@
-import React from 'react';
 import { colors } from '@codecademy/gamut-styles';
+import React from 'react';
+
 import { LogoProCutout } from '../LogoProCutout';
 import { LogoProCutoutTransparent } from '../LogoProCutoutTransparent';