+++ packages/gamut-labs/src/landingPage/Description.tsx
@@ -1,7 +1,8 @@
-import React from 'react';
-import styled from '@emotion/styled';
-
 import { Markdown } from '@codecademy/gamut';
+import styled from '@emotion/styled';
+import React from 'react';
+
+import { BaseProps } from './types';
 
 const DescriptionContainer = styled.div`
   margin: 2rem 0 0;
@@ -12,14 +13,15 @@ export type DescriptionProps = {
   text: string;
   className?: string;
   testId?: string;
-};
+} & Pick<BaseProps, 'onAnchorClick'>;
 
 export const Description: React.FC<DescriptionProps> = ({
   text,
   className,
   testId,
+  onAnchorClick,
 }) => (
   <DescriptionContainer className={className} data-testid={testId}>
-    <Markdown text={text} />
+    <Markdown text={text} onAnchorClick={onAnchorClick} />
   </DescriptionContainer>
 );