+++ packages/gamut-labs/src/landingPage/Feature.tsx
@@ -1,8 +1,9 @@
-import React from 'react';
-import styled from '@emotion/styled';
-
 import { Heading, Markdown } from '@codecademy/gamut';
 import { mediaQueries } from '@codecademy/gamut-styles';
+import styled from '@emotion/styled';
+import React from 'react';
+
+import { BaseProps } from './types';
 
 const Icon = styled.img`
   width: 4rem;
@@ -48,19 +49,7 @@ export type FeatureProps = {
    * Feature image alt text (for screen readers)
    */
   imgAlt?: string;
-
-  /**
-   * Feature title text
-   */
-  title?: string;
-
-  /**
-   * Feature body text as markdown
-   */
-  desc?: string;
-
-  testId?: string;
-};
+} & Pick<BaseProps, 'title' | 'desc' | 'onAnchorClick' | 'testId'>;
 
 export const Feature: React.FC<FeatureProps> = ({
   isIcon,
@@ -68,6 +57,7 @@ export const Feature: React.FC<FeatureProps> = ({
   imgAlt = '',
   title,
   desc,
+  onAnchorClick,
   testId,
 }) => (
   <FeatureBlock data-testid={testId}>
@@ -83,7 +73,7 @@ export const Feature: React.FC<FeatureProps> = ({
     )}
     {desc && (
       <div>
-        <StyledMarkdown text={desc} />
+        <StyledMarkdown text={desc} onAnchorClick={onAnchorClick} />
       </div>
     )}
   </FeatureBlock>