+++ packages/gamut-labs/src/landingPage/PagePrefooter.tsx
@@ -1,9 +1,9 @@
-import React from 'react';
-import styled from '@emotion/styled';
 import { Container } from '@codecademy/gamut';
 import { mediaQueries } from '@codecademy/gamut-styles';
+import styled from '@emotion/styled';
+import React from 'react';
 
-import { CTA, Title, Description } from './';
+import { CTA, Description, Title } from './';
 import { BaseProps } from './types';
 
 const FlexContainer = styled(Container)`
@@ -27,13 +27,16 @@ export const PagePrefooter: React.FC<BaseProps> = ({
   title,
   desc,
   cta,
+  onAnchorClick,
   testId,
 }) => {
   const SectionTitle = title && <Title>{title}</Title>;
-  const Desc = desc && <Description text={desc} />;
+  const Desc = desc && (
+    <Description text={desc} onAnchorClick={onAnchorClick} />
+  );
 
   return cta ? (
-    <FlexContainer nowrap column justify="spaceBetween">
+    <FlexContainer data-testid={testId} nowrap column justify="spaceBetween">
       <FlexContent>
         {SectionTitle}
         {Desc}
@@ -43,7 +46,7 @@ export const PagePrefooter: React.FC<BaseProps> = ({
       </StyledCTA>
     </FlexContainer>
   ) : (
-    <div>
+    <div data-testid={testId}>
       {SectionTitle}
       {Desc}
     </div>